requirements.txt
opencv-python==4.6.0.66
ipython==8.4.0
traitlets==5.3.0
numpy==1.23.0
imutils==0.5.4
easygui==0.98.3
imageio==2.19.3
matplotlib==3.5.2
Pillow==9.1.1
pandas==1.4.3